Fix '#include ...' lines for renamed files
